web_dev/fastapi/simply-social-media-app/pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "fastapi-course-modern"
version = "0.1.0"
description = "Modern rewrite of the FastAPI course project."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "alembic==1.18.4",
  "email-validator>=2.3,<3",
  "fastapi==0.135.3",
  "httpx>=0.28,<0.29",
  "pydantic==2.13.2",
  "pydantic-settings==2.13.1",
  "pwdlib[argon2]>=0.3,<1",
  "psycopg[binary]==3.3.3",
  "pyjwt>=2.10,<3",
  "python-multipart>=0.0.20,<0.1",
  "sqlalchemy==2.0.49",
  "uvicorn==0.43.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4,<9",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
pythonpath = ["."]
testpaths = ["tests"]
